requirements.txt
absl-py==2.1.0
accelerate==0.30.1
afinn==0.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiohttp-retry==2.9.1
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.4.0
astunparse==1.6.3
async-lru==2.0.4
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
basemap==1.3.8
basemap-data==1.3.2
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.7.0
blis==0.7.11
boto3==1.33.8
botocore==1.33.8
branca==0.6.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
cloudpathlib==0.16.0
colorama==0.4.6
comm==0.1.4
confection==0.1.5
contourpy==1.1.0
cvzone==1.6.1
cycler==0.11.0
cymem==2.0.8
debugpy==1.7.0
decorator==5.1.1
defusedxml==0.7.1
executing==1.2.0
fastjsonschema==2.18.0
fastprogress==1.0.3
filelock==3.14.0
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.25
folium==0.14.0
fonttools==4.42.1
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2023.12.1
ftfy==6.3.1
gast==0.5.4
geographiclib==2.0
geopy==2.3.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
greenlet==3.2.3
grpcio==1.64.1
h5py==3.11.0
huggingface-hub==0.23.2
idna==3.4
imageio==2.34.2
intel-openmp==2021.4.0
ipykernel==6.25.2
ipython==8.15.0
ipython-genutils==0.2.0
ipywidgets==8.1.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.0
jieba==0.42.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.4.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.24.0
keras==3.3.3
keras-bert==0.89.0
keras-embed-sim==0.10.0
keras-layer-normalization==0.16.0
keras-multi-head==0.29.0
keras-pos-embd==0.13.0
keras-position-wise-feed-forward==0.8.0
keras-self-attention==0.51.0
keras-transformer==0.40.0
kiwisolver==1.4.5
ktrain==0.41.3
lab==7.4
langcodes==3.4.0
langdetect==1.0.9
language_data==1.2.0
lazy_loader==0.4
libclang==18.1.1
login==0.0.6
marisa-trie==1.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.10.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.1
mkl==2021.4.0
ml-dtypes==0.3.2
mpmath==1.3.0
multidict==6.3.0
murmurhash==1.0.10
mysql==0.0.3
mysql-connector==2.2.9
mysqlclient==2.2.0
namex==0.0.8
narwhals==1.29.0
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.7
networkx==3.2.1
nltk==3.8.1
notebook==7.0.3
notebook_shim==0.2.3
numpy==1.24.1
opencv-python==4.8.1.78
opt-einsum==3.3.0
optree==0.11.0
overrides==7.4.0
packaging==23.1
pandas==2.1.0
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
playsound==1.3.0
preshed==3.0.9
prometheus-client==0.17.1
prompt-toolkit==3.0.39
propcache==0.3.1
protobuf==4.25.3
psutil==5.9.5
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==19.0.1
pycparser==2.21
pydantic==2.7.3
pydantic_core==2.18.4
pydeck==0.9.1
Pygments==2.16.1
PyJWT==2.10.1
pyparsing==3.0.9
pyproj==3.6.1
pyshp==2.3.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3.post1
pywin32==306
pywinpty==2.0.11
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.0
referencing==0.30.2
regex==2024.5.15
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.10.2
s3transfer==0.8.2
safetensors==0.4.3
scikit-image==0.24.0
scikit-learn==1.5.0
scipy==1.11.4
seaborn==0.13.2
Send2Trash==1.8.2
sentencepiece==0.2.0
simplejson==3.19.1
six==1.16.0
smart-open==6.4.0
smmap==5.0.2
sniffio==1.3.0
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.41
srsly==2.4.8
stack-data==0.6.2
streamlit==1.42.2
sympy==1.13.1
syntok==1.4.4
tbb==2021.12.0
tenacity==9.0.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
terminado==0.17.1
textblob==0.18.0.post0
tf_keras==2.16.0
thinc==8.2.3
threadpoolctl==3.5.0
tifffile==2024.7.2
tika==2.6.0
tinycss2==1.2.1
tokenizers==0.19.1
toml==0.10.2
torch==2.6.0+cpu
torchaudio==2.6.0+cpu
torchvision==0.21.0+cpu
tornado==6.3.3
tqdm==4.66.4
traitlets==5.9.0
transformers==4.41.2
twilio==9.5.1
txt2tags==3.8
typer==0.9.4
typing_extensions==4.12.1
tzdata==2023.3
ultralytics==8.3.99
ultralytics-thop==2.0.14
uri-template==1.3.0
urllib3==2.0.4
wasabi==1.1.3
watchdog==6.0.0
wcwidth==0.2.6
weasel==0.3.4
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.3
Werkzeug==3.0.1
Whoosh==2.7.4
widgetsnbextension==4.0.8
wrapt==1.16.0
yarl==1.18.3
gunicorn==22.0.0
python-dotenv==1.0.0
